maintain cleanliness and tidiness

+++ src/demo_arc.py
@@ -1,5 +1,5 @@
-import sys
 import os
+import sys
 import glob
 
 import shlex
@@ -9,19 +9,16 @@ import shutil
 from multiprocessing import Pool
 import multiprocessing as multi
 
-exp_dir = '../exp/arc/arc_dla_34'
-if not os.path.exists(exp_dir):
-    print(f"{exp_dir} doesn't exist")
-    exit(-1)
-
-model = f'{exp_dir}/model_last.pth'
+args = sys.argv
+arch = args[1] # dla_34, hourglass, resdcn_101, resdcn_18
+model = args[2]
 if not os.path.exists(model):
     print(f"{model} doesn't exist")
     exit(-1)
 
-result_dir = f'{exp_dir}/demo'
+result_dir = args[3]
 if not os.path.exists(result_dir):
-    os.mkdir(result_dir)
+    os.makedirs(result_dir, exist_ok=True)
 
 png_dir = '../data/arc/val'
 if not os.path.exists(png_dir):
@@ -29,7 +26,6 @@ if not os.path.exists(png_dir):
     exit(-1)
 
 py_exe = 'python ./demo.py'
-arch = 'dla_34'
 
 def demo(png_path):
     python_script = f"{py_exe} arc --arch {arch} " \
@@ -38,9 +34,12 @@ def demo(png_path):
     png_file = os.path.basename(png_path)
     root, _ = os.path.splitext(png_file)
     txt_file = root + '.txt'
-    shutil.move(png_file, f'{result_dir}/')
-    shutil.move(txt_file, f'{result_dir}/')
+    if os.path.exists(png_file):
+        shutil.move(png_file, f'{result_dir}/')
+    if os.path.exists(txt_file):
+        shutil.move(txt_file, f'{result_dir}/')
 
-p = Pool(multi.cpu_count()//2)
+#p = Pool(multi.cpu_count()//2)
+p = Pool(4)
 p.map(demo, glob.glob(f'{png_dir}/*.png'))
 p.close()
